Changes in the artifact deployment
This chabges the way the deployment to EAP works. At the STI build time
every binary artifact is copied to the /artifacts/ directory. Nothing is
deployed to EAP at that point. The actual deployment is executed at the
boot time of the image. The script waits for the server to come up and
uses the deploy command to deploy all artifacts found in the /artifacts/
directory.